Add code to fix race conditions, add better handling for conneciton termination on graceful shutdown
