+++ remote/server_test.go
@@ -18,7 +18,7 @@ import (
 	"google.golang.org/grpc/credentials"
 )
 
-func createServer(t *testing.T, addr, name string) {
+func createServer(t *testing.T, addr, name string) *Server {
 	certificate, err := tls.LoadX509KeyPair("../test/simple/certs/leaf.cert", "../test/simple/certs/leaf.key")
 	require.NoError(t, err)
 
@@ -54,6 +54,8 @@ func createServer(t *testing.T, addr, name string) {
 		grpcServer.Stop()
 		lis.Close()
 	})
+
+	return s
 }
 
 func startLocalServer(t *testing.T) (string, *string) {
@@ -75,11 +77,15 @@ func startLocalServer(t *testing.T) (string, *string) {
 	return ts.Listener.Addr().String(), &bodyData
 }
 
+var servers []*Server
+
 func setupServers(t *testing.T) (string, *string) {
 	// local server
-	createServer(t, ":1234", "server_local")
-	createServer(t, ":1235", "server_remote_1")
-	createServer(t, ":1236", "server_remote_2")
+	servers = []*Server{
+		createServer(t, ":1234", "server_local"),
+		createServer(t, ":1235", "server_remote_1"),
+		createServer(t, ":1236", "server_remote_2"),
+	}
 
 	// setup the local endpoint
 	return startLocalServer(t)
@@ -137,6 +143,69 @@ func TestExposeRemoteServiceCreatesLocalListener(t *testing.T) {
 	_, err = net.Dial("tcp", "localhost:19000")
 	require.NoError(t, err)
 }
+
+func TestShutdownRemovesLocalListener(t *testing.T) {
+	c, _, _ := setupTests(t)
+
+	resp, err := c.ExposeService(context.Background(), &shipyard.ExposeRequest{
+		Service: &shipyard.Service{
+			Name:                "Test Service",
+			RemoteConnectorAddr: "localhost:1235",
+			SourcePort:          19000,
+			DestinationAddr:     "localhost:19001",
+			Type:                shipyard.ServiceType_REMOTE,
+		},
+	})
+
+	require.NoError(t, err)
+	require.NotEmpty(t, resp.Id)
+
+	time.Sleep(100 * time.Millisecond) // wait for setup
+
+	// check the listener exists
+	_, err = net.Dial("tcp", "localhost:19000")
+	require.NoError(t, err)
+
+	// shutdown
+	for _, s := range servers {
+		s.Shutdown()
+	}
+
+	_, err = net.Dial("tcp", "localhost:19000")
+	require.Error(t, err)
+}
+
+func TestShutdownRemovesRemoteListener(t *testing.T) {
+	c, _, _ := setupTests(t)
+
+	resp, err := c.ExposeService(context.Background(), &shipyard.ExposeRequest{
+		Service: &shipyard.Service{
+			Name:                "Test Service",
+			RemoteConnectorAddr: "localhost:1235",
+			SourcePort:          19000,
+			DestinationAddr:     "localhost:19001",
+			Type:                shipyard.ServiceType_LOCAL,
+		},
+	})
+
+	require.NoError(t, err)
+	require.NotEmpty(t, resp.Id)
+
+	time.Sleep(100 * time.Millisecond) // wait for setup
+
+	// check the listener exists
+	_, err = net.Dial("tcp", "localhost:19000")
+	require.NoError(t, err)
+
+	// shutdown
+	for _, s := range servers {
+		s.Shutdown()
+	}
+
+	_, err = net.Dial("tcp", "localhost:19000")
+	require.Error(t, err)
+}
+
 func TestExposeRemoteServiceCreatesLocalListener2(t *testing.T) {
 	c, _, _ := setupTests(t)
